Add support for reset_password and force_random_password when creating user
